Implement "Singleplayer" networking
The portions of the networking that are required to run the game without any other players is working! The game cycles from Waiting to Syncing to Simulating as expected.

Currently I believe the game is running at 1/2 speed roughly, since it takes 2 ticks to do a full cycle and the simulation time is locked at 16ms